requirements.txt
snowflake-connector-python==3.6.0
PyYAML==6.0.1